[chore] renamed map.tsx to index.tsx
